requirements.txt
requests
python-dotenv
opencv-python
numpy
mss
ultralytics
google-generativeai
fastapi[all]
elevenlabs